get the getActiveCues logic out the engine so the externalCaptionHandler could us it too
