+++ src/engines/html5/html5.js
@@ -6,9 +6,8 @@ import {CustomEventType, Html5EventType} from '../../event/event-type';
 import MediaSourceProvider from './media-source/media-source-provider';
 import VideoTrack from '../../track/video-track';
 import AudioTrack from '../../track/audio-track';
-import PKTextTrack from '../../track/text-track';
+import PKTextTrack, {getActiveCues} from '../../track/text-track';
 import ImageTrack from '../../track/image-track';
-import {Cue} from '../../track/vtt-cue';
 import * as Utils from '../../utils/util';
 import Html5AutoPlayCapability from './capabilities/html5-autoplay';
 import Error from '../../error/error';
@@ -391,7 +390,7 @@ export default class Html5 extends FakeEventTarget implements IEngine {
       this._mediaSourceAdapter.selectTextTrack(textTrack);
     }
     this.resetAllCues();
-    this.addCueChangeListener();
+    this._addCueChangeListener();
   }
 
   /**
@@ -1068,7 +1067,7 @@ export default class Html5 extends FakeEventTarget implements IEngine {
    * @returns {void}
    * @private
    */
-  addCueChangeListener(): void {
+  _addCueChangeListener(): void {
     let textTrackEl = Array.from(this._el.textTracks).find(track => PKTextTrack.isNativeTextTrack(track) && track.mode !== PKTextTrack.MODE.DISABLED);
     if (textTrackEl) {
       this._eventManager.listen(textTrackEl, 'cuechange', (e: FakeEvent) => this._onCueChange(e));
@@ -1093,20 +1092,7 @@ export default class Html5 extends FakeEventTarget implements IEngine {
    * @private
    */
   _onCueChange(e: FakeEvent): void {
-    let textTrack: TextTrack = e.currentTarget;
-    let activeCues: Array<Cue> = [];
-    for (let cue of textTrack.activeCues) {
-      //Normalize cues to be of type of VTT model
-      if (window.VTTCue && cue instanceof window.VTTCue) {
-        activeCues.push(cue);
-      } else if (window.TextTrackCue && cue instanceof window.TextTrackCue) {
-        try {
-          activeCues.push(new Cue(cue.startTime, cue.endTime, cue.text));
-        } catch (error) {
-          new Error(Error.Severity.RECOVERABLE, Error.Category.TEXT, Error.Code.UNABLE_TO_CREATE_TEXT_CUE, error);
-        }
-      }
-    }
+    let activeCues = getActiveCues(e);
     this.dispatchEvent(new FakeEvent(CustomEventType.TEXT_CUE_CHANGED, {cues: activeCues}));
   }
